+++ src/com/runrunfast/homegym/record/RecordDataAction.java
@@ -14,7 +14,7 @@ public class RecordDataAction extends BaseRecordData{
 	public String actionId; // 本次训练动作id
 	public String actionName; // 本次训练动作名称
 	public int groupCount; // 该动作做了几组
-	public float totalKcal; // 该动作消耗的总卡路里
+	public int totalKcal; // 该动作消耗的总卡路里
 	public String actualDate; // 实际完成的日期
 	
 	public List<GroupDetail> groupDetailList = new ArrayList<GroupDetail>();
